auth commit

+++ controllers/student_controllers.js
@@ -3,8 +3,9 @@ let studentsModel = require('../models/students_model')
 exports.createStudent = async (req, res,next) => {
     const {name, age, gen} = req.body;
 
+    const userId = req.user._id;
     try {
-        const student = new studentsModel({name, age, gen})
+        const student = new studentsModel({name, age, gen,userId})
         await student.save();
         res.send({success:true, data:student})
    
@@ -30,8 +31,20 @@ exports.getStudents = async(req, res, next) => {
 
 exports.updateStudents = async (req, res, next) => {
     try {
-        const {name,age,gen} = req.body;
-        const student = await studentsModel.findOneAndUpdate({name:name}, {name:name,age:age,gen:gen});
+        const {name,gen} = req.body;
+        const userId = req.user._id;
+
+        const user = await studentsModel.findOne({name});
+        if(!user){
+            return res.send({success:false, message:"data not found"})
+        }
+
+        if(user.userId.toString() !== userId.toString()){
+            return res.send({success:false, message:"access denied"});
+        }
+
+        const student = await studentsModel.findOneAndUpdate(user._id, {gen});
+
         res.send({success:true, data:student})
     } catch (error) {
         console.log(error)
@@ -44,6 +57,17 @@ exports.deleteStudent = async (req, res, next) => {
 
     try {
         const {name} = req.body;
+        const userId = req.user._id;
+        const user = await studentsModel.findOne({name});
+
+        if(!user){
+            return res.send({success:false, message:"data not found"})
+        }
+
+        if(user.userId.toString() !== userId.toString()){
+            return res.send({success:false, message:"access denied"});
+        }
+
         const student = await studentsModel.findOneAndDelete({name})
         res.send({success:true, data:student})
     } catch (error) {
